requirements.txt
fastapi
uvicorn
python-dotenv
pyvi